requirements.txt
pyautogui==0.9.50
python-dotenv==0.13.0
selenium==3.141.0
